Successfully implemented tree-filter-cps
I have successfully implemented tree-filter-cps however
it is not successfully passing all tests.

Specifically if fails the following tests:
>>> (instrument 'get 'set! '_)
Test $check-no-set
expected 0 obtained 1

>>> (let ((tree (make-node '(aaa) (make-leaf '(b)) (make-leaf '(a))))) (instrument-count-pairs (lambda () (tree-filter-cps member-a tree cons))))
Test $calculate-cons-usage
expected 1 obtained 5

This means a set is being created when it should not be and that
the number of times the continuation is used is too high.